requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-stubs==3.15.1
drf-yasg==1.21.8
greenlet==3.1.1
idna==3.10
inflection==0.5.1
kombu==5.4.2
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
